scripts/pyproject.toml
[tool.poetry]
name = "xmptest"
version = "0.1.0"
description = ""
authors = ["Martin Dullweber <martin@dul.li>"]

[tool.poetry.dependencies]
python = "^3.9"
pypdfium2 = "^4.18.0"
Pillow = "^10.0.0"
numpy = "^1.25.2"
pikepdf = "^8.4.0"

[tool.poetry.dev-dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.284"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
